+++ packages/localizer-simple-number/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.11](https://github.com/jquense/react-widgets/compare/react-widgets-simple-number@5.0.10...react-widgets-simple-number@5.0.11) (2021-11-15)
+
+**Note:** Version bump only for package react-widgets-simple-number
+
+
+
+
+
+## [5.0.10](https://github.com/jquense/react-widgets/compare/react-widgets-simple-number@5.0.9...react-widgets-simple-number@5.0.10) (2021-11-01)
+
+**Note:** Version bump only for package react-widgets-simple-number
+
+
+
+
+
+## [5.0.9](https://github.com/jquense/react-widgets/compare/react-widgets-simple-number@5.0.8...react-widgets-simple-number@5.0.9) (2021-09-21)
+
+**Note:** Version bump only for package react-widgets-simple-number
+
+
+
+
+
 ## [5.0.8](https://github.com/jquense/react-widgets/compare/react-widgets-simple-number@5.0.7...react-widgets-simple-number@5.0.8) (2021-09-07)
 
 **Note:** Version bump only for package react-widgets-simple-number